Remove of StepHelper, all in GuidedtoursHelper
